4.21*: blogilistan laajennus, step10

+++ osa4/blogilista/controllers/blogs.js
@@ -5,7 +5,7 @@ const jwt = require('jsonwebtoken')
 
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -14,12 +14,12 @@ blogsRouter.post('/', async (request, response, next) => {
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
-      return response.status(401).json({ error: 'token missing or invalid' })
+        return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
-    
+
     const blog = new Blog({
-        title:  body.title,
+        title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes,
@@ -28,7 +28,7 @@ blogsRouter.post('/', async (request, response, next) => {
 
     if (blog.likes === null) {
         blog.likes = 0
-    } 
+    }
 
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
@@ -38,21 +38,34 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (decodedToken.id.toString() === blog.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'tried delete blog made by another user'})
+    }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
     const body = request.body
 
     const blog = {
-        title:  body.title,
+        title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes
     }
 
-    await Blog.findByIdAndUpdate(request.params.id, blog, {new : true})
+    await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     response.json(blog.toJSON)
 })
